move tests to server
