+++ packages/datatrak-web-server/src/@types/express/index.d.ts
@@ -1,13 +1,13 @@
-/**
+/*
  * Tupaia
- * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
 
 import { AccessPolicy } from '@tupaia/access-policy';
 import { TupaiaApiClient } from '@tupaia/api-client';
 import { SessionCookie } from '@tupaia/server-boilerplate';
+
 import { DataTrakSessionType, DataTrakSessionModel } from '../../models';
-import { DatatrakWebServerModelRegistry } from '../../types';
 
 declare global {
   namespace Express {
@@ -19,7 +19,6 @@ declare global {
       ctx: {
         services: TupaiaApiClient;
       };
-      models: DatatrakWebServerModelRegistry;
     }
   }
 }
