+++ packages/datatrak-web/src/api/mutations/index.ts
@@ -9,5 +9,4 @@ export { useRegister } from './useRegister';
 export { useEditUser } from './useEditUser';
 export { useResendVerificationEmail } from './useResendVerificationEmail';
 export { useRequestProjectAccess } from './useRequestProjectAccess';
-export { useSubmitSurvey } from './submitSurvey/useSubmitSurvey';
-export { processSurveyResponse } from './submitSurvey/processSurveyResponse';
+export { useSubmitSurvey } from './useSubmitSurvey';
