+++ packages/datatrak-web/src/api/mutations/useSubmitSurvey.ts
@@ -5,13 +5,23 @@
 
 import { useMutation } from 'react-query';
 import { generatePath, useNavigate, useParams } from 'react-router';
-import { Coconut } from '../../../components';
+import { Coconut } from '../../components';
 import { post } from '../../api';
-import { ROUTES } from '../../../constants';
-import { useSurveyForm } from '../../../features';
-import { useSurvey, useUser } from '../../queries';
-import { successToast } from '../../../utils';
-import { processSurveyResponse, AnswersT } from './processSurveyResponse';
+import { ROUTES } from '../../constants';
+import { useSurveyForm } from '../../features';
+import { useSurvey, useUser } from '../queries';
+import { successToast } from '../../utils';
+
+type AutocompleteAnswer = {
+  isNew?: boolean;
+  optionSetId: string;
+  value: string;
+  label: string;
+};
+
+type Answer = string | number | boolean | null | undefined | AutocompleteAnswer;
+
+export type AnswersT = Record<string, Answer>;
 
 // utility hook for getting survey response data
 export const useSurveyResponseData = () => {
@@ -20,11 +30,11 @@ export const useSurveyResponseData = () => {
   const { surveyStartTime, surveyScreenComponents } = useSurveyForm();
   const { data: survey } = useSurvey(surveyCode);
   return {
-    surveyStartTime,
+    startTime: surveyStartTime,
     surveyId: survey?.id,
     questions: Object.values(surveyScreenComponents!).reduce((acc, val) => acc.concat(val), []), // flattened array of survey questions
     countryId: user?.country?.id,
-    useId: user?.id,
+    userId: user?.id,
   };
 };
 
@@ -40,13 +50,8 @@ export const useSubmitSurvey = () => {
         return;
       }
 
-      const processedResponse = processSurveyResponse({
-        ...surveyResponseData,
-        answers,
-      });
-
       await post('submitSurvey', {
-        data: processedResponse,
+        data: { ...surveyResponseData, answers },
       });
     },
     {
